refactor: extract gql schema sidebar content into its own component
