+++ packages/hoppscotch-common/src/components/http/Sidebar.vue
@@ -11,6 +11,7 @@
       :id="'docs'"
       :icon="IconBookOpen"
       :label="`${t('tab.documentation')}`"
+      :order="-2"
     >
       <GqlDocExplorer :key="activeGQLTabId" />
     </HoppSmartTab>
@@ -19,57 +20,9 @@
       :id="'schema'"
       :icon="IconBox"
       :label="`${t('tab.schema')}`"
+      :order="-1"
     >
-      <div
-        v-if="schemaString"
-        class="sticky top-0 z-10 flex flex-shrink-0 items-center justify-between overflow-x-auto border-b border-dividerLight bg-primary pl-4"
-      >
-        <label class="truncate font-semibold text-secondaryLight">
-          {{ t("graphql.schema") }}
-        </label>
-        <div class="flex">
-          <HoppButtonSecondary
-            v-tippy="{ theme: 'tooltip' }"
-            to="https://docs.hoppscotch.io/documentation/protocols/graphql"
-            blank
-            :title="t('app.wiki')"
-            :icon="IconHelpCircle"
-          />
-          <HoppButtonSecondary
-            v-tippy="{ theme: 'tooltip' }"
-            :title="t('state.linewrap')"
-            :class="{ '!text-accent': WRAP_LINES }"
-            :icon="IconWrapText"
-            @click.prevent="toggleNestedSetting('WRAP_LINES', 'graphqlSchema')"
-          />
-          <HoppButtonSecondary
-            v-tippy="{ theme: 'tooltip' }"
-            :title="t('action.download_file')"
-            :icon="downloadSchemaIcon"
-            @click="downloadSchema"
-          />
-          <HoppButtonSecondary
-            v-tippy="{ theme: 'tooltip' }"
-            :title="t('action.copy')"
-            :icon="copySchemaIcon"
-            @click="copySchema"
-          />
-        </div>
-      </div>
-      <div
-        v-if="schemaString"
-        :key="activeGQLTabId"
-        class="h-full relative w-full"
-      >
-        <div ref="schemaEditor" class="absolute inset-0"></div>
-      </div>
-      <HoppSmartPlaceholder
-        v-else
-        :src="`/images/states/${colorMode.value}/blockchain.svg`"
-        :alt="`${t('empty.schema')}`"
-        :text="t('empty.schema')"
-      >
-      </HoppSmartPlaceholder>
+      <GqlSchema :key="activeGQLTabId" />
     </HoppSmartTab>
 
     <!-- Standard REST sidebar tabs -->
@@ -140,44 +93,28 @@
 </template>
 
 <script setup lang="ts">
-import { useCodemirror } from "@composables/codemirror"
 import IconBookOpen from "~icons/lucide/book-open"
 import IconBox from "~icons/lucide/box"
-import IconCheck from "~icons/lucide/check"
 import IconClock from "~icons/lucide/clock"
-import IconCopy from "~icons/lucide/copy"
-import IconDownload from "~icons/lucide/download"
-import IconHelpCircle from "~icons/lucide/help-circle"
 import IconLayers from "~icons/lucide/layers"
 import IconFolder from "~icons/lucide/folder"
 import IconShare2 from "~icons/lucide/share-2"
 import IconCode from "~icons/lucide/code"
 import IconServer from "~icons/lucide/server"
-import IconWrapText from "~icons/lucide/wrap-text"
-import { computed, reactive, ref, watch } from "vue"
+import { computed, ref, watch } from "vue"
 import { useI18n } from "@composables/i18n"
-import { useColorMode } from "@composables/theming"
-import { useToast } from "@composables/toast"
-import { copyToClipboard } from "@helpers/utils/clipboard"
-import { refAutoReset } from "@vueuse/core"
 import { useService } from "dioc/vue"
 import MockServerDashboard from "~/components/mockServer/MockServerDashboard.vue"
-import { useNestedSetting } from "~/composables/settings"
 import { useMockServerWorkspaceSync } from "~/composables/mockServerWorkspace"
 import { useMockServerVisibility } from "~/composables/mockServerVisibility"
 import { GQLTabConnectionService } from "~/services/gql-tab-connection.service"
-import { toggleNestedSetting } from "~/newstore/settings"
-import { platform } from "~/platform"
 import { WorkspaceTabsService } from "~/services/tab/workspace-tabs"
 
 const t = useI18n()
-const colorMode = useColorMode()
-const toast = useToast()
 
 const tabs = useService(WorkspaceTabsService)
 const gqlTabConn = useService(GQLTabConnectionService)
 
-const schemaString = gqlTabConn.activeTabSchemaString
 const activeGQLTabId = gqlTabConn.activeGQLTabId
 
 const { isMockServerVisible } = useMockServerVisibility()
@@ -212,76 +149,6 @@ watch(isGQLTab, (nowGQL) => {
   }
 })
 
-// GQL Schema sidebar
-const downloadSchemaIcon = refAutoReset<typeof IconDownload | typeof IconCheck>(
-  IconDownload,
-  1000
-)
-const copySchemaIcon = refAutoReset<typeof IconCopy | typeof IconCheck>(
-  IconCopy,
-  1000
-)
-
-const schemaEditor = ref<any | null>(null)
-const WRAP_LINES = useNestedSetting("WRAP_LINES", "graphqlSchema")
-
-useCodemirror(
-  schemaEditor,
-  schemaString,
-  reactive({
-    extendedEditorConfig: {
-      mode: "graphql",
-      readOnly: true,
-      lineWrapping: WRAP_LINES,
-    },
-    linter: null,
-    completer: null,
-    environmentHighlights: false,
-  })
-)
-
-const downloadSchema = async () => {
-  const dataToWrite = schemaString.value
-  const file = new Blob([dataToWrite], { type: "application/graphql" })
-  const url = URL.createObjectURL(file)
-
-  const filename = `${
-    url.split("/").pop()!.split("#")[0].split("?")[0]
-  }.graphql`
-
-  URL.revokeObjectURL(url)
-
-  const result = await platform.kernelIO.saveFileWithDialog({
-    data: dataToWrite,
-    contentType: "application/graphql",
-    suggestedFilename: filename,
-    filters: [
-      {
-        name: "GraphQL Schema File",
-        extensions: ["graphql"],
-      },
-    ],
-  })
-
-  if (result.type === "unknown" || result.type === "saved") {
-    downloadSchemaIcon.value = IconCheck
-    toast.success(`${t("state.download_started")}`)
-  }
-}
-
-const copySchema = () => {
-  if (!schemaString.value) return
-
-  copyToClipboard(schemaString.value)
-  copySchemaIcon.value = IconCheck
-}
-
 // Ensure mock servers are kept in sync with workspace changes globally
 useMockServerWorkspaceSync()
 </script>
-
-<style lang="scss" scoped>
-:deep(.cm-panels) {
-  @apply top-sidebarPrimaryStickyFold #{!important};
-}
-</style>
